HyphaFile: Remove locking from read method
Now that files are never written directly, only through a temporary file
that atomically overwrites the original file, it is no longer needed to
lock files for reading. Once a reader opens the file for reading, it can
read that file undisturbed, even when it is replaced by an update (since
that creates a new file, without touching the file being read).

This relates to #265.