Add HyphaDatatypePage::getSortDateTime() method
This allows datatypes to define a custom sorting datetime, for use in
e.g. a future tag index page. Ideally, this would always return the
"publish" date, but only peer_reviewed_article really defines this. In
absence of that, the last modified date is used.

For mailinglist and festival, there is no clear sort date, so these just
return null for now.