tagindex: Work with special characters in tag names
This includes spaces, which are urlencoded by the browser when making a
request. To allow these, the tag label must be urldecoded before using
it.

Ideally, this would happen centrally (see #274), but for now just do
this here.

Note that until urls are also urlencoded on generation, this might still
break for some special characters in tag labels (notably /). This should
be fixed as part of #274.