Show errors in wymeditor image upload plugin
The upload handler would report upload errors by returning a call to an
`uploadResponse` javascript function (this is possible because the
result of this handler is loaded into an iframe).

Originally, the `uploadResponse` function was responsible for handling
both successful uploads and errors. However, in commit 75b2242
(wymeditor: Updated WYMeditor), the custom image upload was replaced by
the standard wymeditor image upload, so then successful uploads were
handled by returning JSON instead.

Unfortunately, that plugin does not seem to handle errors at all, which
is probably why the old (but broken) `uploadResponse` call was left in
there.

This commit makes the backend return errors as a proper JSON object, and
modifies the wymeditor plugin to handle these (by simply showing an
alert popup).

This fixes #234.