Move index system pages into datatypes
Previously, the various types of indexes were distinguished by the
loadPage function which generated the pagename and let other functions
handle the main html.

To make this more consistent with other system pages (settings,
tagindex), create an indexpage system datatype for these indices
instead.

This commit only moves code, it does not change the generated output or
the way it is generated.