settingspage: Move authentication check into class
Previously, the authentication check happened in loadPage, while the
rest of the page was rendered by the settingspage class. Other datatypes
handle auth checks inside the class, which is cleaner and keeps loadPage
shorter.

This commit applies the same for settingspage, moving the check into
settingspage::process.

When the user is not logged in, the user is redirected to the homepage
instead of rendering the settingspage. Because this now happens in the
process function, it can just return a 'redirect' command instead of
setting the Location header directly and aborting further processing
with exit. Additionally, this just redirects to the root url, rather
than the default page directly, since that resolves to the default page
anyway.