indexpage: Fix language selection
Originally, the first element in the url path was interpreted as
the selected language for all pages. Since quite some time, url parsing
was refactored and this language parsing only happens for datatype
pages, no longer for system pages.

However, the index still assumed the old behaviour:
 - Alternative index languages linked to e.g. `/en/index`,
   resulting in a page not found error.
 - The rendering code used the current contentlanguage to decide which
   index to render, which is never set to anything but the default
   language for system pages.

This fixes both problems:
 - Links are now generated as `/index/en`.
 - The rendering code now gets the language to use from the url itself
   (falling back to the content language if no language was given or it
   is invalid).

This fixes part of #310