RequestContext: Store CSRF token in a cookie
Previously, it was stored in the session. Since every page contains a
(login) form, this means that every visitor needs to have a session.

By storing the token in a cookie instead, the number of sessions can be
reduced (this does require some changes in the session handling still).

Because setcookie does not update $_COOKIE directly, this adds a
`$csrfToken` attribute on RequestContext to keep the current value,
loading it from $_COOKIE only in the constructor.