RequestContext: Merge Hypha class into RequestContext
The Hypha class used to collect a number of HyphaFile instances and the
currently selected theme, but it was never really used for much else.
Later, RequestContext was introduced which is actually a better place
for all this.

Also, because the Hypha class was statically initialized very early
(when base.php was loaded), HyphaRequest and RequestContext were not
instantiated yet, so this blocked any use of HyphaRequest and
RequestContext when initializing it. By delaying the initialization to
the RequestContext constructor, this limitation is lifted.

This also moves the code that decides what theme to use into
RequestContext, replaces the `data->theme` variable with a
`RequestContext::getThemeName()` and renames `data->html` and
`data->css` to `data->themeHtml` and `data->themeCss` respectively, to
better reflect their purpose.