HyphaSession: Add class
This class wraps all access to the PHP session, with an API similar to
HyphaFile and Xml. Collection all session accesses makes it easier to
apply some improvements next.

This new class is always accessed through HyphaRequest, which reduces
the use of global variables again.

This commit should mostly move code. The only functional changes are:
 - The session is now first initialized when the HyphaRequest is
   created, so a bit later than before.
 - The way notifications are stored in the session is changed slightly,
   to make it easier to use the get/set API of HyphaSession.