HyphaSession: Comment that session writes on load are intentional
Whenever loading the session, it is also written. There was a comment
that said this was only because there was no way to close the session
without writing before PHP5.6, but it turns out that writing to the
session on every request is actually useful to keep the session alive
(since the session garbage collector seems to look at the session
*modified* time, rather than the *access* time. This updates the comment
to indicate the writing is actually intentional.