Fix cancel button on "forgot password" popup
This used to call `showLogin()` to show the login popup (since you
typically get to this popup from the login popup). However, the login
popup is shown using `login()`, rather than `showLogin()`.

This was already broken in the first git commit, so this probably never
worked, or got broken in a refactor way before that.

This fixes #305.