dump: Simplify a bit
This removes the call_user_func_array, which was no longer needed. It
was used in an earlier version of this feature to pass the `$vars` array
as separate arguments to another PHP function, but this was later
changed to just generating a complete `console.log` line directly, so
this can just call `writeScript` normally instead.

Also, no point to check for an empty list, foreach on an empty list will
simply not do anything.