dump: convert values to JSON early
In PHP, objects are passed by reference rather than by value. This means
that if you would dump an object, and then change it, the changed value
would be displayed. This did not happen for most other values (strings,
integers, arrays) since those are passed by value (that is, a copy would
be stored in the `$hyphaDumpList` array).

This commit changes dump to convert any dumped values to JSON right
away, so the value printed is the value when dump is called and excludes
any changes afterwards.