dump: Improve display of iterables and DomNodes
When dumping a DomNode, it is serialized to XML, and deserialized to a
Javascript DomNode again, which can be conveniently explored in the
browser console.

When dumping an iterable (other than an array), dump the elements
separately. This is particularly useful for things like NodeList, which
is otherwise just serialized to an empty object.

For the DomNode conversion to work, the resulting expression is no
longer valid JSON (since it includes Javascript code). This could not be
done with json_encode, so this adds a `convertForDump()` function that
essentially re-implements json_encode partly. This needs reimplements the
array encoding and traversal, so that any DomNodes (even when buried in
a nested array or NodeList) can return raw javascript expressions
(rather than being encoded into a JSON string). For non-array values,
json_encode is still used.