peer_reviewed_article: Pass single HyphaDomElement to addDiscussion
Previously, this would pass a NodeList containing (normally) just a
single element. However, a NodeList does not seem to have a `nodeName`
property, which `addDiscussion` used to differentiate between review and
public comments.

This caused `addDiscussion` treat review discussions as public
discussions, leaving out the `user` attribute. This in turn prevented
users from resolving their own blocking comments, since there was no
longer a record of the initiator of a discussion (#318).

Passing a single HyphaDomElement rather than a NodeList to
`addDiscussion()` makes the latter work as expected again.

This fixes #318.