HyphaMacros: Add class to support macros in HTML
This registers a post-processing function that looks for any
`<macro name="abc">` tags in the final HTML. The named macro will be
invoked, which can then replace the macro tag with other HTML.

This also adds a trivial "now" macro, which simply writes the current
timestamp with a configurable format.

Macros also get passed a context, which is always empty now. In the
future, datatypes (or loadPage) could explicitly process macros earlier
in rendering, allowing macros to use e.g. the current page as context.

When any errors occur processing the macro, an error is logged and the
macro tag is replaced by a generic error message.