Fix POST url for newPage
In commit 1fea30d (Refactor newpage popup HTML generation), a / was
accidentally removed from the POST URL for newPage, resulting in urls
like `/enpagename`.

In practice, this did not prevent creating a page, since the language
and pagename were also sent as POST data and the subsequent redirect
used that data.

However, when creating a page that already existed (or when another
error occured), a reload was used, which would result in the invalid url
being requested and the rather cryptic combination of "page alrady
exists" and "Page doesn't exist" notifications.