wymeditor: Support the <macro> tag
This changes the WymEditor parser and validator to allow the macro tag
to be used.

It also adds it to WymEditors ROOT_CONTAINERS, BLOCKS and
NON_CONTAINING_ELEMENTS lists, in an attempt to make editing around
these macros a bit more reliable (which is not perfect yet, it seems the
browser or WymEditor is still prone to remove or duplicate the macro if
you edit close to it.

Finally, this addds some CSS to make the macro visible in the editor.

This modifies the wymeditor sources. The minified version is
regenerated using the `uglify` command (`apt-get install uglify`).

    uglifyjs jquery.wymeditor.js --comments > jquery.wymeditor.min.js