PageListMacro: Add macro
This macro can be used to generate a list of pages, optionally filtered
in various ways and sorted.

The code is heavily based on tagindexpage and they should probably share
code at some point, but for now this just duplicates some of the code
(with slight modifications).

Some CSS is added for the page list too, which mostly just reuses the
styling for the tagindex (except that this shows a separating line above
and below the list too, which is not done on the tagindex).