festival: Explicitly convert string time to int
This code takes the hours value out of a time string and does
calculations with it. This relies on PHP to implicitly convert the
substring to an int, which is fragile and confuses type checking. This
makes the conversion explicit, and uses named variables for slightly
more readable code too.