festival: Remove useless implode/explode sequences
There is no need to encode this data as a string, just leave it as an
array. This is simpler, more robust and less prone to type check
problems (i.e. it does not rely on implicit string-to-integer
conversion).

This also allows removing now useless zero padding in the column numbers
(this was needed to ensure proper sorting of the encoded strings, it
seems).