htaccess: Do not enable display_errors php flag
Enabling error display is generally not a good idea on production
servers, so instead of enabling this unconditionally just rely on the
server default configuration. Also, php_flag does not work in all
webserver configurations (e.g. not when using PHP-FPM).

If needed, display errors can still be enabled by creating a "DEBUG"
file in the repository root directory.

This fixes #327.