WIP: festival: Allow translating signup and contribution form
This no longer reads the forms from the config elements directly, but
looks for a language subtag with the right id.

Note that this does not allow the contributions themselves to be
translated yet, just the displayed form.

Also note that this breaks compatibility, without some form of
migration, any custom forms will become unused.