wymeditor: Adds paste listener to strip superfluous code
Note that this makes *all* pastes text-only and also prevents
copy-pasting within the same editor, or between editors. This should be
solved later.

There also seems to be a corner case in the `wym.paste` function where
the text is interpreted as HTML rather than text (i.e. any HTML tags in
the text content, which must have been escaped as e.g. `&gt;` in the
original clipboard contents, would become actual HTML again after the
paste). This only occurs when inserting into the outer body tag, so when
no other container is selected (which is pretty hard to achieve, except
when the editor is completely empty). This also seems to occur with the
existing "Paste from word" feature as well. Since this is an existing
and rare problem, it is left to be solved later.

This fixes the main part of #66