peer_reviewed_article: Do not notify from updateToNewStatus()
This method should just modify the state and return whether it worked or
not, leaving the user notification to the actual action that calls it.

This method is also called to set the status to STATUS_APPROVED after
an approve action or block resolve, which do not really need this notify
either (if the status change would be invalid, this would be a bug, not
a user error).