peer_reviewed_article: Add review checklist
In the DRAFT, REVIEW and APPROVED stages, this shows a checklist of
things to do before progressing to the next stage, which helps to
quickly see if anything is forgotten.

Some of these things are marked as required, meaning they prevent
progress (i.e. starting review or publishing). Others are marked as
recommended, meaning they are shown but not enforced. Whether something
is required or recommended is hardcoded for now, but the code is
writting to allow making this configurable later.

This commit just shows the checklist and hides the state transition
buttons if some required checks fail, enforcing these checks in the
backend will be added in a separate commit.

This fixes #304.