peer_reviewed_article: Enforce checks in status change handling
This ensures that all status transitions are prevented when the needed
checks are not satisfied.

The buttons for these transitions were already hidden in the frontend,
this just enforces the same things in the backend.

Note that this updates the `statusChangeAction()`, not
`updateToNewStatus()`, since the latter is also used for the REVIEW ->
APPROVED transition, which should not be prevented when other checks
fail (e.g. consider the case where all tags are removed after starting
review, the third approval is added, the has-tags-check prevents the
status from moving to APPROVED, then the tags are restored, but that
does not (and can not easily) trigger a transition to APPROVED).