timestampToDateTime: Add helper function
This converts a UNIX timestamp to a datetime value, while handling empty
timestamps and also setting the timezone to the default timezone (which
is useful when formatting them for display or comparing just the date
part).

This helper is used right away in the getSortDateTime() method of
peer_reviewed_article and text.