mailinglist: Display progress for send
Set intermediate progress updates in the xml.
Lock and unload browser page with javascript and show a modal with the progress until all mails have been send. This uses a newly defined json  endpoint for progress retrieval. And an extra parameter to the  makeAction function with a javascript callback upon completion.

Review fixes