HTMLForm: Forbid <script> tags in wymeditor fields
The WymEditor already strips these client-side, but this could be
circumvented by users or code running on users' computers (i.e.
viruses). This ensures that wymeditor fields are not considered valid
when they contain <script> tags. This check is done by letting
DomDocument parse the HTML, to maximize the chance of catching
malevolent script tags, even when they are obfuscated.

This fixes #358.