PagelistMacro: Always filter by current language
Showing pages in other languages did not work reliably due to
limitations elsewhere. Until those are lifted, better disable custom
language filtering and always filter by the current language instead.
There is one exception: Filtering by just the current language is
allowed, to allow a seamless transition if this limitation is lifted
later and the default is changed to not filter by language.