base: Add tags to body class attribute
This adds both the tag id (which is stable between languages and when
tag label editing becomes possible) and the tag label (which is easier
to read in the CSS). The label version is "slugified" using
validatePageName, which seems to off a reasonable set of rules for
generating a CSS class too.

This fixes #381