RequestContext: Make CSRF-cookie https-only again
In commit 4d1cfd3 (Use different cookie names on HTTPS #368) a
non-existing `$secure` variable was used instead of the previous call to
`isSecure()`. This was probably a leftover from an earlier
implementation that was left in the final commit. Since the variable did
not exist, this caused a notice and caused the CSRF cookie to never be
marked as https-only.

This reverts that part of that commit, making the CSRF cookie https-only
again (for https requests).